Mark failed tests with pytest.xfail
